29/04/2023 - Carlos - Refactorización de código sacando las funciones que interactuan con el POD a un fichero a parte
