+++ webapp/src/components/pod/ButtonAddPod.tsx
@@ -1,20 +1,16 @@
 import "../../css/navigation.css";
 import {useSession} from "@inrupt/solid-ui-react";
-import {getFile, overwriteFile} from "@inrupt/solid-client";
 import {Button, Container} from "@mui/material";
 import React, {useState} from "react";
-import {Session} from "@inrupt/solid-client-authn-browser";
 import Notification from "../Notification";
 import ReactDOM from "react-dom/client";
 import MapView from "../map/MapView";
 import Icon from "../../img/symbols/GOMapSymbol.png";
 import {initReactI18next, useTranslation} from "react-i18next";
 import i18n from "../../i18n";
-import FriendList from "./FriendList";
 import Filter from "../map/options/Filter";
-
-import {v4 as uuidv4} from "uuid";
 import ImgbbUploader from "../map/ImgbbUploader";
+import {createData, createMarker} from "./PODsInteraction";
 
 
 i18n.use(initReactI18next)
@@ -35,99 +31,8 @@ function ButtonAddPod({idName, idCategory, idComment, idScore, idLatitude, idLon
     const {t} = useTranslation();
     const [error, setError] = useState(false)
     const [noSelectedMap, setNoSelectedMap] = useState(false)
-
-    async function createMarker(idName: string, idCategory: string, idComment: string, idScore: string,
-                                idLatitude: string, idLongitude: string, fileURL: string)  {
-        let name = (document.getElementById(idName) as HTMLInputElement).value;
-        let identifier = fileURL.split("lomap")[0] + "profile/card#me"
-        let category = (document.getElementById(
-            idCategory
-        ) as HTMLInputElement).value;
-        let comment = (document.getElementById(
-            idComment
-        ) as HTMLInputElement).value;
-        let latitude = (document.getElementById(
-            idLatitude
-        ) as HTMLInputElement).value;
-        let longitude = (document.getElementById(
-            idLongitude
-        ) as HTMLInputElement).value;
-
-        let imgUrl = (document.getElementById(
-            "upload-img"
-        ) as HTMLInputElement)?.src;
-
-        let json = {
-            "@context": "https://schema.org/",
-            "@type": "Place",
-            "identifier":uuidv4(),
-            "name": name,
-            "author": {
-                "@type":"Person",
-                "identifier": identifier
-            },
-            "additionalType": category,
-            "latitude": latitude,
-            "longitude": longitude,
-            "description": comment,
-            "review": [],
-            "image": [{
-                "@type": "ImageObject",
-                "author": {
-                    "@type": "Person",
-                    "identifier": identifier
-                },
-                "contentUrl": imgUrl
-            }],
-            "dateCreated": new Date().valueOf()
-        };
-
-        if (imgUrl === null || imgUrl === undefined) {
-            json.image = [];
-        }
-
-        return await readFileFromPod(fileURL, session).then(file => {
-                if (file === "") {
-                    let fileContent = [json]
-                    const blob = new Blob([JSON.stringify(fileContent, null, 2)], {
-                        type: "application/ld+json",
-                    });
-                    return new File([blob], fileURL, {type: blob.type});
-                } else {
-                    let fileContent = JSON.parse(file);
-                    fileContent.spatialCoverage.push(json);
-                    const blob = new Blob([JSON.stringify(fileContent, null, 2)], {
-                        type: "application/ld+json",
-                    });
-                    return new File([blob], fileURL, {type: blob.type});
-                }
-            }
-        );
-    };
-
-    async function readFileFromPod(fileURL: string, session: Session){
-        try {
-            const file = await getFile(
-                fileURL,
-                {fetch: session.fetch}
-            );
-            return file.text();
-        } catch (err) {
-            return "";
-        }
-    }
-
-    async function createData(url: string, file: File, session: Session){
-        try {
-            await overwriteFile(
-                url,
-                file,
-                {contentType: file.type, fetch: session.fetch}
-            );
-        } catch (error) {
-            console.log(error);
-        }
-    };
+    const [showNotification, setShowNotification] = useState(false);
+    const [imageUrl, setImageUrl] = useState("");
 
     const handleCloseNotification = () => {
         setShowNotification(false);
@@ -135,7 +40,7 @@ function ButtonAddPod({idName, idCategory, idComment, idScore, idLatitude, idLon
         setNoSelectedMap(false)
     };
 
-    const [showNotification, setShowNotification] = useState(false);
+
     const createNotification = () => {
         setShowNotification(true);
         setTimeout(() => {
@@ -156,7 +61,7 @@ function ButtonAddPod({idName, idCategory, idComment, idScore, idLatitude, idLon
             if(route===""){
                 setNoSelectedMap(true)
             }else{
-                createMarker(idName, idCategory, idComment, idScore, idLatitude, idLongitude, route)
+                createMarker(idName, idCategory, idComment, idScore, idLatitude, idLongitude, route,session)
                     .then((file) => createData(route, file, session))
                     .then(createNotification)
                     .then(() => {
@@ -189,8 +94,6 @@ function ButtonAddPod({idName, idCategory, idComment, idScore, idLatitude, idLon
             }
         }
     };
-
-    const [imageUrl, setImageUrl] = useState("");
 
     function handleUploadSuccess(imageUrl: string) {
         setImageUrl(imageUrl);
